requirements.txt
black==25.1.0
Django==4.2.23
django-environ>=0.11
django-filter==25.1
django-simple-history==3.10.1
flake8==7.3.0
graphene-django==3.2.3
psycopg2-binary==2.9.10
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
requests==2.32.4
